pyproject.toml
[project]
name = "rag-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "chromadb>=1.5.7",
    "fastapi>=0.135.3",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-text-splitters>=0.3.0",
    "langchain-groq>=1.1.2",
    "pypdf>=6.10.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "sentence-transformers>=5.4.1",
    "uvicorn>=0.44.0",
    "langchain-huggingface>=1.2.1",
    "langchain-chroma>=1.1.0",
]
